몬스터 스포너 적용 & 몬스터 10종 제작 & 스테이지 구성

+++ BrawlStars/Assets/Scripts/Item/DropTable.cs
@@ -19,6 +19,9 @@ public class DropTable : ScriptableObject
 
 	public DropItem[] GetDropItemList(string monsterName)
 	{
-		return dropTable[monsterName];
+		if (dropTable.ContainsKey(monsterName))
+			return dropTable[monsterName];
+		else
+			return null;
 	}
 }
